+++ CHANGELOG.md
@@ -6,6 +6,8 @@
  - Search field can now be reset with escape
 ### Fixed
  - User name input not shown in sharing tab
+ - Import ignores client side encryption settings
+ - Backup does not reset some user settings
 
 ## 2019.8.1 - Bugfix release
 ### Changed